Create Cloudwatch 1
In this project, I created a CloudWatch dashboard in AWS to monitor the performance of an EC2 instance. The dashboard shows key metrics like CPU utilization and network activity (in and out). This helps in identifying any performance issues, errors, or unusual activity in the instance. Such monitoring is useful for maintaining the health and security of AWS resources.